requirements.txt
fastapi
uvicorn
face_recognition
numpy
opencv-python
pillow
python-multipart